Kernel+Lib: moved `heap_allocator` as a library and kernel depends on it.
This is a move to make this library used in user space, since the method used in both sides is very similar.

We get new pages either from the virtual memory or `inc_heap` syscall, and then add it to the free pages.

Is this the best allocator implementation? of course not XD
Is it working? yes, maybe?
Is it fun? YES
