+++ Cargo.toml
@@ -4,5 +4,5 @@ default-members = ["kernel"]
 members = [
     "libraries/kernel_user_link",
     "kernel",
-    "userspace/init", "userspace/shell",
+    "userspace/init", "userspace/shell", "libraries/increasing_heap_allocator",
 ]
