+++ libraries/increasing_heap_allocator/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "kernel"
+name = "increasing_heap_allocator"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-kernel_user_link = { path = "../libraries/kernel_user_link" }
-increasing_heap_allocator = { path = "../libraries/increasing_heap_allocator" }
